+++ material/src/plugins/metadataSelector/dialogMetadataSelector/dialogMetadataSelector.interface.ts
@@ -4,32 +4,51 @@ import {GridMetadata, MetadataSelector, MetadataSelectorOptions, VisualPluginOpt
 /**
  * Css classes for dialog metadata selector
  */
-export interface CssClassesDialogMetadataSelector
+export interface CssClassesDialogMetadataSelector<TCssClasses = unknown>
 {
-    componentClass: string;
-    btnClass: string;
-    btnIconClass: string;
-    dialogComponentClasses: Object;
+    /**
+     * Class for show metadata button element
+     */
+    btnElement: string;
+
+    /**
+     * Class for show metadata button icon element
+     */
+    btnIconElement: string;
+
+    /**
+     * Css classes for component displayed as dialog
+     */
+    dialogComponentClasses: TCssClasses;
 }
 
 /**
  * Texts that are used within DialogMetadataSelector
  */
-export interface DialogMetadataSelectorTexts
+export interface DialogMetadataSelectorTexts<TTexts = unknown>
 {
+    /**
+     * Text displayed for button that displayes metadata selector
+     */
     btnShowSelection: string;
-    dialogComponentTexts: Object;
+
+    /**
+     * Object storing texts for dialog component
+     */
+    dialogComponentTexts: TTexts;
 }
 
 /**
  * Dialog metadata selector options
  */
-export interface DialogMetadataSelectorOptions<TMetadata = unknown> extends MetadataSelectorOptions, VisualPluginOptions<CssClassesDialogMetadataSelector>
+export interface DialogMetadataSelectorOptions<TMetadata extends GridMetadata = GridMetadata,
+                                               TCssClasses = unknown,
+                                               TTexts = unknown> extends MetadataSelectorOptions, VisualPluginOptions<CssClassesDialogMetadataSelector<TCssClasses>>
 {
     /**
      * Texts that are used within DialogMetadataSelector
      */
-    texts: DialogMetadataSelectorTexts;
+    texts: DialogMetadataSelectorTexts<TTexts>;
 
     /**
      * Indication whether is button for showing metadata selection visible
@@ -39,25 +58,27 @@ export interface DialogMetadataSelectorOptions<TMetadata = unknown> extends Meta
     /**
      * Name of storage storing current metadata status
      */
-    storageName: string;
+    storageName: string|undefined|null;
 
     /**
      * Component that is used for handling metadata selection itself
      */
-    dialogComponent: Type<DialogMetadataSelectorContentComponent<TMetadata>>;
+    dialogComponent: Type<DialogMetadataSelectorContentComponent<TMetadata, TCssClasses, TTexts>>;
 }
 
 /**
  * Public API for 'DialogMetadataSelector'
  */
-export interface DialogMetadataSelector<TMetadata extends GridMetadata = GridMetadata> extends MetadataSelector<TMetadata>
+export interface DialogMetadataSelector<TMetadata extends GridMetadata = GridMetadata> extends MetadataSelector<TMetadata, DialogMetadataSelectorOptions<TMetadata>>
 {
 }
 
 /**
  * Data that are passed to component that handles metadata
  */
-export interface DialogMetadataSelectorComponentData<TMetadata = unknown>
+export interface DialogMetadataSelectorComponentData<TMetadata extends GridMetadata = GridMetadata,
+                                                     TCssClasses = unknown,
+                                                     TTexts = unknown>
 {
     /**
      * Method that is used for setting metadata into component
@@ -67,12 +88,12 @@ export interface DialogMetadataSelectorComponentData<TMetadata = unknown>
     /**
      * Css classes passed to dialog component
      */
-    cssClasses: Object;
+    cssClasses: TCssClasses;
 
     /**
      * Texts passed to dialog component
      */
-    texts: Object;
+    texts: TTexts;
 
     /**
      * Method that is used for sending metadata out of component to metadata selector
@@ -84,10 +105,12 @@ export interface DialogMetadataSelectorComponentData<TMetadata = unknown>
 /**
  * Component that is rendered within dialog
  */
-export interface DialogMetadataSelectorContentComponent<TMetadata = unknown>
+export interface DialogMetadataSelectorContentComponent<TMetadata extends GridMetadata = GridMetadata,
+                                                        TCssClasses = unknown,
+                                                        TTexts = unknown>
 {
     /**
      * Data that are used for communication with MetadataSelector
      */
-    data: DialogMetadataSelectorComponentData<TMetadata>;
+    data: DialogMetadataSelectorComponentData<TMetadata, TCssClasses, TTexts>;
 }