+++ changelog.md
@@ -427,6 +427,8 @@
 - updated `CssClassesPreviousNextPaging` interface, completely refactored, incompatible with previous one
 - updated `CssClassesLoadMorePaging` interface, completely refactored, incompatible with previous one
 - updated `CssClassesSimpleNoDataRenderer` interface, completely refactored, incompatible with previous one
+- updated `CssClassesVerticalDragNDropSelection` interface, completely refactored, incompatible with previous one
+- updated `CssClassesDialogMetadataSelector` interface, completely refactored, incompatible with previous one
 - updated all paging plugins to correspond with new `Paging` interface
 - updated all plugins to be in line with new `GridPlugin` interface
 - *subpackage* `@anglr/grid/extensions`